[iOS Simulator] http/tests/preload/viewport/meta-viewport-link-headers.php is a flaky failure. https://bugs.webkit.org/show_bug.cgi?id=172518
This unflakes the test by downloading a "timer" image and kicking off the timer that runs the
test only once that image was downloaded, while potentially waiting for the preloaded images to arrive,
up to 1 second.
That means that slow test environments would have a longer timer than today, avoiding flakiness.

Reviewed by Youenn Fablet.

* http/tests/preload/viewport/meta-viewport-link-headers.php: Add a "timer" image, which load event starts the test.
* platform/ios/TestExpectations: Unflake the test in expectations.


Identifier: 192975@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@221604 268f45cc-cd09-0410-ab3c-d52691b4dbfc