REGRESSION (r215784): The title of right-to-left pages are empty https://bugs.webkit.org/show_bug.cgi?id=176746 <rdar://problem/34211419>
Reviewed by Brent Fulgham.

Source/WebCore:

Left truncate a long right-to-left title.

Right-to-left text represents the visual ordering of text. Internally WebKit stores
right-to-left text identically to left-to-right text in memory. So, we can use the
same string operation to truncate right-to-left text as we do to truncate left-to-right
text.

* platform/text/StringWithDirection.h:
(WebCore::truncateFromEnd):

Tools:

Add unit test to ensure we do not regress StringWithDirection.truncateFromEnd().

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebCore/StringWithDirection.cpp: Added.
(TestWebKitAPI::TEST):

Identifier: 193262@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@221921 268f45cc-cd09-0410-ab3c-d52691b4dbfc