[GTK][Wayland] Flickering when resizing the window https://bugs.webkit.org/show_bug.cgi?id=176774
Reviewed by Michael Catanzaro.

This regressed when we switched to always start the wayland compositor. In
AcceleratedDrawingAreaProxy::waitForAndDispatchDidUpdateBackingStoreState() we check if the nested compositor is
running, which is always true now (if EGL and the required extensions are available). We should check if we are
in AC mode instead.

* UIProcess/AcceleratedDrawingAreaProxy.cpp:
(WebKit::AcceleratedDrawingAreaProxy::waitForAndDispatchDidUpdateBackingStoreState):

Identifier: 193358@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@222027 268f45cc-cd09-0410-ab3c-d52691b4dbfc